Bug fix CE button
Before, CE button always erased last operation made on lascCalcHistory. When the last operation was an operator, it would erase it and the user could put a number, and that would create a bug where we could put two operations using numbers simultaniously.

Now this operation checks if the last calculation history is a number before clearing it out.